sasha_agent/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
pyyaml==6.0.1
google-generativeai==0.3.2
python-multipart==0.0.9
aiohttp==3.9.3
typing-extensions==4.9.0 